% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GF_twomode_function.R
\name{GF_twomode}
\alias{GF_twomode}
\title{Gould-Fernandez Two-Mode Brokerage}
\usage{
GF_twomode(g, event_attribute, actor_attribute)
}
\arguments{
\item{g}{igraph object.}

\item{event_attribute}{This needs to be a dataframe with the firm name and the sector attribute}

\item{actor_attribute}{This needs to a dataframe with the person name and the gender attribute}
}
\value{
List - 1.)Brokerage path list 2.) Individual role dataframe list
3.) Total two-mode GF results dataframe.
}
\description{
This function extracts the two mode brokerage chains from a two-mode network. Only works for interlocking directorates
}
\references{
Jasny, L. and Lubell, M., 2015. Two-mode brokerage in policy networks. Social Networks, 41, pp.36-47.
}
